fix: update tests/to_unicode_tests.cpp

+++ tests/to_unicode_tests.cpp
@@ -29,7 +29,7 @@ std::string read_file(std::string filename) {
   while (stream.read(&buf[0], read_size)) {
     out.append(buf, 0, size_t(stream.gcount()));
   }
-  out.append(buf, 0, stream.gcount());
+  out.append(buf, 0, size_t(stream.gcount()));
   return out;
 }
